Remove static_cast from errno comparisons

+++ src/socket.cpp
@@ -81,7 +81,7 @@ void Socket::connect(const char _addr[], const int _port) const
 	// case.
 	// CHECK if this is the right way to complete this TODO
 
-	if (res == -1 && errno != static_cast<int>(EINPROGRESS) ) {
+	if (res == -1 && errno != EINPROGRESS ) {
 		throw std::runtime_error(net::method::getErrorMsg(errno));
 	} else if (res == 0) {
 		throw std::invalid_argument("Address argument invalid");
@@ -98,7 +98,7 @@ Socket Socket::accept() const
 
 	auto client = ::accept(sockfd, (sockaddr *) peerAddr.get(), &peerAddrSize);
 
-	if (client == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (client == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for client == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -113,7 +113,7 @@ void Socket::write(const std::string &_msg) const
 {
 	const auto written = low_write(::write, sockfd, _msg);
 
-	if (written == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (written == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for written == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -127,7 +127,7 @@ void Socket::send(const std::string &_msg, SF::send _flags) const
 	const auto flags = static_cast<int>(_flags);
 	const auto sent  = low_write(::send, sockfd, _msg, flags);
 
-	if (sent == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (sent == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for sent == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -143,7 +143,7 @@ void Socket::send(
 	const auto sent  = low_write(
 	  ::sendto, sockfd, _msg, flags, (sockaddr *) &_addr, sizeof(_addr));
 
-	if (sent == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (sent == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for sent == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -159,7 +159,7 @@ std::string Socket::read(const int _bufSize) const
 
 	const auto recvd = low_read(::read, sockfd, str);
 
-	if (recvd == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (recvd == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for bytes == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -177,7 +177,7 @@ std::string Socket::recv(const int _bufSize, SF::recv _flags) const
 	const auto flags = static_cast<int>(_flags);
 	const auto recvd = low_read(::recv, sockfd, str, flags);
 
-	if (recvd == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (recvd == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for recvd == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
@@ -199,7 +199,7 @@ std::string Socket::recv(
 	const auto recvd
 	  = low_read(::recvfrom, sockfd, str, flags, (sockaddr *) &_addr, &length);
 
-	if (recvd == -1 && errno != static_cast<int>(EAGAIN) && errno != static_cast<int>(EWOULDBLOCK)) {
+	if (recvd == -1 && errno != EAGAIN && errno != EWOULDBLOCK) {
 		// TODO: Check for recvd == EAGAIN or EWOULDBLOCK and
 		// don't throw exception in that case.
 		// CHECK if this is the right way to complete this TODO
